requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django>=1.11.5
django-braces==1.13.0
django-crispy-forms==1.7.2
dj-database-url==0.5.0
django-debug-toolbar==1.10.1
django-embed-video==1.1.2
django-redis==4.9.0
djangorestframework==3.8.2
idna==2.7
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
redis==2.10.6
requests>=2.19.1
sqlparse==0.2.4
urllib3==1.23
django-heroku==0.3.1
gunicorn==19.9.0
whitenoise==4.1.2
